requirements.txt
asgiref==3.5.2
astroid==2.12.12
certifi==2022.6.15.1
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.6
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.50.0
django-auto-logout==0.5.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-mysql==4.7.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
eradicate==2.1.0
filelock==3.8.0
future==0.18.2
identify==2.5.8
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
mando==0.6.4
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
mysqlclient==2.1.1
nodeenv==1.7.0
oauthlib==3.2.0
openapi-codec==1.3.2
Pillow==9.3.0
pipenv==2022.10.12
platformdirs==2.5.2
pre-commit==2.20.0
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==1.7.1
pylama==8.4.1
pylint==2.15.4
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
radon==5.1.0
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.6
virtualenv-clone==0.5.7
vulture==2.6
whitenoise==6.2.0
wrapt==1.14.1
